+++ src/contrib/net/sf/samtools/SAMBinaryTagAndValue.java
@@ -60,7 +60,7 @@ public class SAMBinaryTagAndValue {
 
     @Override
     public int hashCode() {
-        int result = (int) tag;
+        int result = tag;
         result = 31 * result + value.hashCode();
         return result;
     }
